Consolodate tests
