+++ tests/lib.rs
@@ -76,7 +76,5 @@ mod subsequents;
 mod tempdir;
 mod undefined_variables;
 #[cfg(target_family = "windows")]
-mod windows_powershell;
-#[cfg(target_family = "windows")]
 mod windows_shell;
 mod working_directory;
