--- tests/windows_powershell.rs
@@ -1,18 +1,0 @@
-use super::*;
-
-#[test]
-fn windows_poweshell_setting_uses_powershell() {
-  Test::new()
-    .justfile(
-      r#"
-      set windows-powershell
-
-      foo:
-        Write-Output bar
-    "#,
-    )
-    .shell(false)
-    .stdout("bar\r\n")
-    .stderr("Write-Output bar\n")
-    .run();
-}
