+++ tests/windows_shell.rs
@@ -35,3 +35,20 @@ fn windows_powershell_setting_uses_powershell() {
     .stderr("Write-Output bar\n")
     .run();
 }
+
+#[test]
+fn windows_poweshell_setting_uses_powershell() {
+  Test::new()
+    .justfile(
+      r#"
+      set windows-powershell
+
+      foo:
+        Write-Output bar
+    "#,
+    )
+    .shell(false)
+    .stdout("bar\r\n")
+    .stderr("Write-Output bar\n")
+    .run();
+}
